requirements.txt
Pillow
opencv-python
pdf2image
svgpathtools
numpy
